@@ -1,9 +1,9 @@
-name: qadritravels
+name: alutabus
 description: A Bus Ticket Booking app
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -30,20 +30,21 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.5
   timeline_list: ^0.0.6
-  font_awesome_flutter: ^10.2.1
-  flutter_offline: ^2.1.0
-  onesignal_flutter: ^3.4.2
-  firebase_auth: ^4.0.2
+  font_awesome_flutter: ^10.5.0
+  flutter_offline: ^3.0.0
+  onesignal_flutter: ^3.5.1
+  firebase_auth: ^4.7.2
   firebase_core:
-  google_sign_in: ^5.4.2
-  simple_animations: ^5.0.0+2
-  intl: ^0.17.0
-  get: any
+  cloud_firestore:
+  google_sign_in: ^6.1.4
+  google_maps_flutter: ^2.4.0
+  simple_animations: ^5.0.2
+  intl: ^0.18.1
+  get: ^4.6.5
 
 dev_dependencies:
   flutter_test:
@@ -54,14 +55,13 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^2.0.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
