@@ -5,8 +5,8 @@ const bermudaGray = Color(0xFF728BAB);
 const mystic = Color(0xFFDFE5ED);
 const heather = Color(0xFFBAC4D4);
 const cadetBlue = Color(0xFFAFB9CA);
-const radicalRed = Color(0xFFF42E60);
+const radicalGreen = Color(0xFF50C878);
 const smoky = Color(0xFF5F5471);
-const kobi = Color(0xFFE4A4BF);
+const kobi = Color.fromARGB(255, 179, 231, 197);
 const backgroundColor = Color(0xFFFCFDFF);
 const darkAccent = Color(0xFF455F82);
