@@ -1,13 +1,13 @@
 List<List<int>> defaultSeats = [
-  [1, 1, 0, 0, 1, 1, 1],
-  [1, 1, 0, 0, 1, 1, 1],
-  [1, 1, 0, 0, 1, 1, 1],
-  [1, 1, 0, 0, 1, 1, 2],
-  [1, 1, 0, 0, 1, 1, 1],
-  [1, 1, 0, 0, 2, 2, 2],
-  [1, 1, 0, 0, 1, 1, 1],
-  [1, 1, 0, 0, 1, 1, 1],
-  [1, 1, 0, 0, 1, 1, 1],
+  [0, 0, 0, 0, 0, 1, 1],
+  [0, 0, 0, 2, 2, 1, 1],
+  [0, 0, 0, 1, 1, 1, 1],
+  // [0, 0, 0, 0, 1, 1, 2],
+  // [1, 1, 0, 0, 1, 1, 1],
+  // [1, 1, 0, 0, 2, 2, 2],
+  // [1, 1, 0, 0, 1, 1, 1],
+  // [1, 1, 0, 0, 1, 1, 1],
+  // [1, 1, 0, 0, 1, 1, 1],
 ];
 
 Map<String, List<List<int>>> seat = {
