@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-import 'package:qadritravels/screens/search_buses.dart';
-import 'package:qadritravels/themes/colors.dart';
+import 'package:alutabus/screens/search_buses.dart';
+import 'package:alutabus/themes/colors.dart';
+import 'package:get/get.dart';
 
 class CustomBottomSearch extends StatelessWidget {
   final TextEditingController? fromController;
@@ -17,83 +18,139 @@ class CustomBottomSearch extends StatelessWidget {
       this.color})
       : super(key: key);
 
+  static const locations = [
+    'Samaru',
+    'Congo',
+    'ABUTH',
+    'Site II',
+    'IAR',
+    'Gym',
+    'Faculty of Physical Science'
+  ];
+
   @override
   Widget build(BuildContext context) {
-    String from, destination;
+    String from = '', destination = '';
 
     return Card(
       clipBehavior: Clip.antiAlias,
       color: color,
       margin: EdgeInsets.zero,
-      shape: const RoundedRectangleBorder(
-          borderRadius: BorderRadius.only(topRight: Radius.circular(42))),
       elevation: elevation,
       child: Padding(
-        padding: const EdgeInsets.only(
-          bottom: 10,
-          right: 10,
-          left: 10,
-          top: 10,
-        ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            TextField(
-              controller: fromController,
-              style: const TextStyle(fontSize: 22.0, color: radicalRed),
-              decoration: const InputDecoration(
-                  icon: Icon(
-                    FontAwesomeIcons.circleDot,
-                    color: radicalRed,
+        padding: const EdgeInsets.all(20),
+        child: Row(
+          children: [
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: <Widget>[
+                  TextField(
+                    controller: fromController,
+                    keyboardType: TextInputType.none,
+                    onTap: () => showModalBottomSheet(
+                        context: context,
+                        builder: (_) => Container(
+                              padding: const EdgeInsets.all(20),
+                              child: ListView.builder(
+                                  itemCount: locations.length,
+                                  itemBuilder: (_, index) => ListTile(
+                                        title: Text(locations[index]),
+                                        onTap: () {
+                                          fromController!.text =
+                                              locations[index];
+                                          Navigator.of(_).pop();
+                                        },
+                                      )),
+                            )),
+                    style: const TextStyle(fontSize: 14.0, color: radicalGreen),
+                    decoration: const InputDecoration(
+                        icon: Icon(
+                          FontAwesomeIcons.circleDot,
+                          color: radicalGreen,
+                        ),
+                        labelText: 'From',
+                        border: InputBorder.none,
+                        labelStyle: TextStyle(color: bermudaGray)),
                   ),
-                  labelText: 'From',
-                  border: InputBorder.none,
-                  labelStyle: TextStyle(color: bermudaGray)),
-            ),
-            Row(
-              crossAxisAlignment: CrossAxisAlignment.center,
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: <Widget>[
-                const Flexible(
-                  child: Padding(
-                    padding: EdgeInsets.only(left: 38.0),
+                  const Padding(
+                    padding: EdgeInsets.only(left: 38),
                     child: Divider(
                       color: bermudaGray,
                     ),
                   ),
-                ),
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: FloatingActionButton(
-                    backgroundColor: darkAccent,
-                    onPressed: () {
-                      from = fromController!.text;
-                      destination = destinationController!.text;
-                      Navigator.push(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => SearchBuses(
-                                    from: from,
-                                    destination: destination,
-                                  )));
-                    },
-                    child: const Icon(Icons.search),
+                  TextField(
+                    controller: destinationController,
+                    keyboardType: TextInputType.none,
+                    onTap: () => showModalBottomSheet(
+                        context: context,
+                        builder: (_) => Container(
+                              padding: const EdgeInsets.all(20),
+                              child: ListView.builder(
+                                  itemCount: locations.length,
+                                  itemBuilder: (_, index) => ListTile(
+                                        title: Text(locations[index]),
+                                        onTap: () {
+                                          destinationController!.text =
+                                              locations[index];
+                                          Navigator.of(_).pop();
+                                        },
+                                      )),
+                            )),
+                    style: const TextStyle(
+                        fontSize: 14.0, color: Colors.deepPurple),
+                    decoration: const InputDecoration(
+                        icon: Icon(
+                          FontAwesomeIcons.circleDot,
+                          color: Colors.deepPurple,
+                        ),
+                        border: InputBorder.none,
+                        labelText: 'Destination',
+                        labelStyle: TextStyle(color: bermudaGray)),
                   ),
-                ),
-              ],
+                ],
+              ),
             ),
-            TextField(
-              controller: destinationController,
-              style: const TextStyle(fontSize: 22.0, color: Colors.deepPurple),
-              decoration: const InputDecoration(
-                  icon: Icon(
-                    FontAwesomeIcons.circleDot,
-                    color: Colors.deepPurple,
-                  ),
-                  border: InputBorder.none,
-                  labelText: 'Destination',
-                  labelStyle: TextStyle(color: bermudaGray)),
+            Align(
+              alignment: Alignment.centerRight,
+              child: FloatingActionButton(
+                backgroundColor: Get.theme.colorScheme.primary,
+                onPressed: () {
+                  if (fromController == null || destinationController == null) {
+                    Get.showSnackbar(const GetSnackBar(
+                        message: 'input controllers not found!'));
+
+                    return;
+                  }
+                  try {
+                    from = fromController!.text;
+                    destination = destinationController!.text;
+                    if (from.isEmpty || destination.isEmpty) {
+                      Get.showSnackbar(const GetSnackBar(
+                          message: 'origin or destination is missing!'));
+                      return;
+                    }
+                    debugPrint('$from - $destination');
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => SearchBuses(
+                                  from: from,
+                                  destination: destination,
+                                )));
+                  } on Exception catch (e) {
+                    e.printError();
+                    Get.showSnackbar(GetSnackBar(
+                        message:
+                            'failed to search a bus for route $from-$destination'));
+                  }
+                },
+                child: Icon(
+                  Icons.search,
+                  color: Get.theme.colorScheme.onPrimary,
+                ),
+              ),
             ),
           ],
         ),
